backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
firebase-admin==6.4.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
onnxruntime==1.16.3
joblib==1.3.2
websockets==12.0
aiofiles==23.2.1
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2
